structures some code
