+++ backend/src/main/java/org/rdfarchitect/models/cim/relations/model/CIMResourceTypeIdentifyingUtils.java
@@ -84,10 +84,4 @@ public class CIMResourceTypeIdentifyingUtils {
         return types.getFirst().getObject().asResource()
                     .hasProperty(CIMS.stereotype, CIMStereotypes.enumeration);
     }
-
-    public boolean isExternalResource(Resource resource) {
-        return !resource.listProperties()
-                    .filterDrop(stmt -> stmt.getPredicate().equals(RDFA.uuid))
-                    .hasNext();
-    }
 }
