+++ backend/src/main/java/org/rdfarchitect/services/delete/FindDeleteDependenciesUseCase.java
@@ -22,7 +22,13 @@ import org.rdfarchitect.database.GraphIdentifier;
 
 import java.util.UUID;
 
-public interface FindOnDeleteRelationsUseCase {
+public interface FindDeleteDependenciesUseCase {
 
-    AffectedResource getDeleteRelations(GraphIdentifier graphIdentifier, UUID uuid);
+    /**
+     * Finds the resources that would be affected by deleting a specified resource.
+     * @param graphIdentifier The identifier of the graph.
+     * @param uuid The resource to find dependencies for.
+     * @return An {@link AffectedResource} containing the affected resources with and their relations.
+     */
+    AffectedResource getDeleteDependencies(GraphIdentifier graphIdentifier, UUID uuid);
 }
